==> 01_wireless_earbuds_data_cleaning.ipynb <==
import pandas as pd

df = pd.read_csv('new_reference_data_for_beats.csv')

print("First five rows of the dataset:")
print(df.head())

print("\nDataFrame summary:")
print(df.info())

print(df.isnull().sum())
print(df.duplicated().sum())

print(df["product_id"].value_counts())
print(df["product_id"].unique())

product_mapping = {
    'B096SV8SJG': 'Beats Studio Buds',
    'B08ZR5JB9G': 'Sennheiser Consumer Audio CS',
    'B09HN594TL': 'Jlab Go Air Pop',
    'B0C2F5KD26': 'Skullcandy Smokin Buds',
    'B09FM6PDHP': 'JBL Tune 230NC',
    'B093SLWMS7': 'Soundcore by Anker Life P3',
    'B0BYPFNW6T': 'Sony WF-C700N',
    'B0BTZKP1TP': 'Jabra Elite 4',
    'B0CF7GYNW2': 'Samsung Galady Buds FE',
    'B07PXGQC1Q': 'Apple AirPods'
}

df['product_name'] = df['product_id'].map(product_mapping)

print(df[['product_id', 'product_name']].drop_duplicates())

df.to_csv("earbuds_reviews_cleaned.csv", index=False)
