==> 03_eda_section1.ipynb <==
# 03_eda_section1

import pandas as pd
import numpy as np

df = pd.read_csv("earbuds_reviews_prepared_for_analysis.csv")

print("First five rows of dataset: ")
print(df.head())

print("\nDataFrame summary:")
print(df.info())

print("\nMissing values per column:")
print(df.isnull().sum())

print("\nNumber of duplicate rows:")
print(df.duplicated().sum())

print("\nReview count by product:")
print(df["product_name"].value_counts())

print("\nReview count by rating:")
print(df["rating"].value_counts().sort_index())

print("\nRating Statistics")
print("Mean =", df["rating"].mean())
print("Median =", df["rating"].median())
print("Mode =", df["rating"].mode()[0])
print("Variance =", df["rating"].var())
print("Standard Deviation =", df["rating"].std())
print("Min =", df["rating"].min())
print("Max =", df["rating"].max())

print("\nReview Length Statistics")
print("Mean =", df["review_length"].mean())
print("Median =", df["review_length"].median())
print("Variance =", df["review_length"].var())
print("Standard Deviation =", df["review_length"].std())
print("Min =", df["review_length"].min())
print("Max =", df["review_length"].max())

print("\nHelpful Count Statistics")
print("Mean =", df["helpful_count"].mean())
print("Median =", df["helpful_count"].median())
print("Variance =", df["helpful_count"].var())
print("Standard Deviation =", df["helpful_count"].std())
print("Min =", df["helpful_count"].min())
print("Max =", df["helpful_count"].max())

Q1 = df["review_length"].quantile(0.25)
Q3 = df["review_length"].quantile(0.75)
IQR = Q3 - Q1

lower = Q1 - 1.5 * IQR
upper = Q3 + 1.5 * IQR

outliers = df[(df["review_length"] < lower) | (df["review_length"] > upper)]

print("\nOutliers (review_length using IQR):")
print(outliers.shape[0])

df.to_csv("eda_part1_summary_dataset.csv", index=False)
